+++ src/Modules/Login/LoginGateway.php
@@ -58,6 +58,11 @@ class LoginGateway extends BaseGateway
         return $fsid;
     }
 
+    public function getLastLogin(int $fsId): string
+    {
+        return $this->db->fetchValueByCriteria('fs_foodsaver', 'last_login', ['id' => $fsId]);
+    }
+
     public function isActivated(int $fsId): bool
     {
         $isActivated = $this->db->fetchValueByCriteria('fs_foodsaver', 'active', ['id' => $fsId]);