+++ tests/Unit/SessionMayRoleTest.php
@@ -9,9 +9,7 @@ use Foodsharing\Lib\Db\Mem;
 use Foodsharing\Lib\Session;
 use Foodsharing\Modules\Core\DBConstants\Foodsaver\Role;
 use Foodsharing\Modules\Foodsaver\FoodsaverGateway;
-use Foodsharing\Modules\Login\LoginGateway;
 use Foodsharing\Modules\Region\RegionGateway;
-use Foodsharing\Modules\Settings\SettingsGateway;
 use Tests\Support\UnitTester;
 
 class SessionTestAdapter extends Session
@@ -20,11 +18,9 @@ class SessionTestAdapter extends Session
         private Mem $mem,
         private FoodsaverGateway $foodsaverGateway,
         private RegionGateway $regionGateway,
-        private LoginGateway $loginGateway,
-        private SettingsGateway $settingsGateway,
         protected bool $initialized = false
     ) {
-        parent::__construct($mem, $foodsaverGateway, $regionGateway, $loginGateway, $initialized);
+        parent::__construct($mem, $foodsaverGateway, $regionGateway, $initialized);
     }
 
     public function setTestUser(Role $role)
@@ -45,8 +41,6 @@ class SessionMayRoleTest extends Unit
             $this->tester->get(Mem::class),
             $this->tester->get(FoodsaverGateway::class),
             $this->tester->get(RegionGateway::class),
-            $this->tester->get(LoginGateway::class),
-            $this->tester->get(SettingsGateway::class),
             true
         );
     }